pyproject.toml
[tool.poetry]
name = "FastDTLmapper"
version = "0.1.0"
description = "Fast genome-wide DTL(Duplication-Transfer-Loss) event mapping tool"
authors = ["moshi"]

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
biopython = "^1.79"
scipy = "^1.7.1"
goatools = "^1.1.6"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.9b0"
pydocstyle = "^6.1.1"
ipykernel = "^6.4.1"
ipython = "^7.27.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
